+++ backend/package.json
@@ -4,8 +4,9 @@
   "description": "Backend for Purehealth Profit Management System",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "prekill": "npx kill-port 5000",
+    "start": "npm run prekill && node server.js",
+    "dev": "npm run prekill && nodemon server.js"
   },
   "dependencies": {
     "@simplewebauthn/server": "^7.4.0",
