+++ frontend/docu/Branch_Logs.md
@@ -526,6 +526,8 @@ feature/Build-Settings-Revamp
 
 feature/feature-Monthly-Income-Export-Revamp
 
+devpatch12
+
