+++ frontend/src/pages/Transaction.jsx
@@ -1,38 +1,32 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Sidebar from '../components/dashboard/Sidebar';
-import Income from '../assets/icons/income_logo.png';
-import Expense from '../assets/icons/expense_logo.png';
 import { Download } from 'lucide-react';
 import useAuth from '../hooks/auth/useAuth';
 import useProtectedAction from '../hooks/auth/useProtectedAction';
 import WebAuthModal from '../components/auth/WebAuthModal';
 import Loading from '../components/transaction/Loading';
-import { useQueryClient, useQuery } from '@tanstack/react-query';
-import { ToastContainer, toast } from 'react-toastify'; // Added toast import
+import { useQueryClient } from '@tanstack/react-query';
+import { ToastContainer, toast } from 'react-toastify';
 import { calculateRefundTotal } from '../utils/transactionUtils';
 import DateSelector from '../components/transaction/DateSelector';
 import IncomeTable from '../components/transaction/IncomeTable';
-import ExpenseTable from '../components/transaction/ExpenseTable';
 import ConfirmationModal from '../components/transaction/ConfirmationModal';
 import TransactionSummaryModal from '../components/transaction/TransactionSummaryModal';
-import ExpenseSummaryModal from '../components/transaction/ExpenseSummaryModal';
 import { useTransactionManagement } from '../hooks/transaction/useTransactionManagement';
 import { useTransactionData } from '../hooks/transaction/useTransactionData';
-import { expenseAPI, settingsAPI } from '../services/api';
 import { exportIncomeToExcel } from '../utils/incomeExcelExporter';
-import { exportExpenseToExcel } from '../utils/expenseExcelExporter';
+import { settingsAPI } from '../services/api';
+import { useQuery } from '@tanstack/react-query';
 
-const Transaction = () => {
+const NewTransaction = () => {
   const { user, isAuthenticating } = useAuth();
   const [selectedDate, setSelectedDate] = useState(new Date());
   const [expenseDate, setExpenseDate] = useState(new Date()); 
   const [searchTerm, setSearchTerm] = useState('');
-  const [expenseSearchTerm, setExpenseSearchTerm] = useState('');
   const [pendingRefundAmount, setPendingRefundAmount] = useState(0);
-  const [isExpenseSummaryOpen, setIsExpenseSummaryOpen] = useState(false);
-  const [selectedExpense, setSelectedExpense] = useState(null);
-  const [isEditingExpense, setIsEditingExpense] = useState(false);
+  const [currentPage, setCurrentPage] = useState(1);
+  const [itemsPerPage] = useState(10);
 
   // WebAuthn protected actions hook
   const {
@@ -46,19 +40,7 @@ const Transaction = () => {
     clearError
   } = useProtectedAction();
   
-  const idTypeOptions = [
-    { value: 'Regular', label: 'Regular' },
-    { value: 'Person with Disability', label: 'PWD' },
-    { value: 'Senior Citizen', label: 'Senior Citizen' }
-  ];
-  
-  const incomeDateInputRef = useRef(null);
-  const expenseDateInputRef = useRef(null);
-  
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-
-  // Fetch discount categories FIRST (needed by useTransactionData)
+  // Fetch discount categories
   const {
     data: discountCategoriesData
   } = useQuery({
@@ -74,8 +56,18 @@ const Transaction = () => {
 
   // Get active discount categories
   const discountCategories = discountCategoriesData?.data?.categories?.filter(cat => cat.status === 'active') || [];
+  
+  // Build idTypeOptions dynamically
+  const idTypeOptions = [
+    { value: 'Regular', label: 'Regular' },
+    ...(discountCategories.map(cat => ({ value: cat.categoryName, label: cat.categoryName })))
+  ];
+  
+  const incomeDateInputRef = useRef(null);
+  const navigate = useNavigate();
+  const queryClient = useQueryClient();
 
-  // Use our new custom hook to handle data fetching and processing
+  // Use our custom hook to handle data fetching and processing
   const {
     transactions,
     departments,
@@ -97,20 +89,7 @@ const Transaction = () => {
     departmentTotals,
     departmentBalanceTotals,
     departmentRefundTotals
-  } = useTransactionData(selectedDate, expenseDate, discountCategories);
-
-  // Fetch categories for the expense modal
-  const { data: categoriesData } = useQuery({
-    queryKey: ['categories'],
-    queryFn: () => expenseAPI.getAllCategories().then(res => res.data),
-    onError: (error) => console.error('Failed to fetch categories:', error),
-  });
-
-  // Process categories data consistently
-  const categories = categoriesData ? 
-    (Array.isArray(categoriesData) ? categoriesData : 
-    (categoriesData.data && Array.isArray(categoriesData.data) ? categoriesData.data : [])) 
-    : [];
+  } = useTransactionData(selectedDate, expenseDate);
 
   // Use the existing transaction management hook
   const {
@@ -176,15 +155,17 @@ const Transaction = () => {
   const departmentsWithValues = getDepartmentsWithValues();
   const { totalGross, totalGCash } = calculateTotalValues(filteredTransactions);
 
-  // Filter expenses by search term
-  const filteredExpenses = filterExpenses(expenses, expenseSearchTerm);
-  const totalExpense = calculateTotalExpense(filteredExpenses);
-
   // Calculate refund total
   const refundInfo = calculateRefundTotal(filteredTransactions);
   const totalRefundAmount = refundInfo.totalRefundAmount;
   const exceededRefunds = Object.values(refundAmounts || {}).reduce((sum, val) => sum + (parseFloat(val) || 0), 0);
   const totalRefundsToDisplay = totalRefundAmount + exceededRefunds + (pendingRefundAmount || 0);
+
+  // Calculate deposit (On-hand cash)
+  const depositAmount = Math.max(0, totalGross - totalRefundsToDisplay - totalGCash);
+
+  // Calculate total amount
+  const totalAmount = totalGross;
 
   const handleDateChange = (e) => {
     const newDate = new Date(e.target.value);
@@ -195,22 +176,6 @@ const Transaction = () => {
     }
   };
 
-  const handleExpenseDateChange = (e) => {
-    const newDate = new Date(e.target.value);
-    if (!isNaN(newDate.getTime())) {
-      const normalizedDate = new Date(
-        newDate.getFullYear(),
-        newDate.getMonth(),
-        newDate.getDate()
-      );
-      
-      setExpenseDate(normalizedDate);
-      setExpenseSearchTerm(''); 
-      refetchExpenseData(normalizedDate);
-    }
-  };
-  
-  const handleNewExpenses = () => navigate('/add-expenses');
   const handleNewIncome = () => navigate('/add-transaction');
 
   // Handle Excel export
@@ -232,141 +197,30 @@ const Transaction = () => {
     }
   };
 
-  // Handle Expense Excel export
-  const handleGenerateExpenseReport = async () => {
-    try {
-      await exportExpenseToExcel(
-        filteredExpenses,
-        totalExpense,
-        expenseDate
-      );
-      toast.success('Expense report exported successfully!');
-    } catch (error) {
-      console.error('Expense export failed:', error);
-      toast.error('Failed to export expense report. Please try again.');
-    }
-  };
-
-  const [localExpenseMenuId, setLocalExpenseMenuId] = useState(null);
-
-  const toggleExpenseMenuFixed = (id) => {
-    
-    if (localExpenseMenuId === id) {
-      setLocalExpenseMenuId(null);
-    } else {
-      setLocalExpenseMenuId(id);
-    }
-  };
-
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      // Only run this if a menu is open
-      if (localExpenseMenuId && !event.target.closest('.menu-dropdown') && !event.target.closest('.menu-button')) {
-        setLocalExpenseMenuId(null);
-      }
-    };
-
-    document.addEventListener('click', handleClickOutside);
-    return () => {
-      document.removeEventListener('click', handleClickOutside);
-    };
-  }, [localExpenseMenuId]);
-
   useEffect(() => {
     setSearchTerm('');
+    setCurrentPage(1);
   }, [selectedDate]);
 
-  const handleEditExpense = (expense) => {
-    setSelectedExpense(expense);
-    setIsEditingExpense(true);
-    setIsExpenseSummaryOpen(true);
-  };
-  
-  const getExpenseModalData = () => {
-    if (!selectedExpense) return {};
-    
-    const firstName = selectedExpense.firstName || selectedExpense.name?.split(' ')[0] || '';
-    const lastName = selectedExpense.lastName || selectedExpense.name?.split(' ').slice(1).join(' ') || '';
-    
-    const expenses = selectedExpense.ExpenseItems && selectedExpense.ExpenseItems.length > 0 
-      ? selectedExpense.ExpenseItems.map((item, index) => ({
-          id: item.id || index,
-          paidTo: item.paidTo || firstName + ' ' + lastName,
-          purpose: item.purpose || '',
-          categoryId: item.Category?.categoryId || item.categoryId || '',
-          categoryName: item.Category?.name || 'No Category',
-          amount: parseFloat(item.amount || 0)
-        }))
-      : [{
-          id: 1,
-          paidTo: firstName + ' ' + lastName,
-          purpose: selectedExpense.purpose || selectedExpense.description || '',
-          categoryId: selectedExpense.Category?.categoryId || selectedExpense.categoryId || '',
-          categoryName: selectedExpense.Category?.name || 'No Category', 
-          amount: parseFloat(selectedExpense.amount || 0)
-        }];
-    
-    return {
-      firstName,
-      lastName,
-      selectedDepartment: selectedExpense.departmentId || '',
-      selectedDate: expenseDate.toISOString().split('T')[0],
-      expenses
+  // Close menu when clicking outside
+  useEffect(() => {
+    if (openMenuId === null) return;
+
+    const handleClickOutside = (event) => {
+      // Close menu if clicking outside
+      toggleIncomeMenu(openMenuId);
     };
-  };
-  
-  const calculateExpenseTotal = () => {
-    const modalData = getExpenseModalData();
-    return modalData.expenses ? modalData.expenses.reduce((total, exp) => total + exp.amount, 0) : 0;
-  };
-  
-  const closeExpenseSummary = () => {
-    setIsExpenseSummaryOpen(false);
-    setSelectedExpense(null);
-    setIsEditingExpense(false);
-  };
 
-  const handleSaveExpenseChanges = async (updatedData) => {
-    try {
-      const totalAmount = updatedData.expenses.reduce((sum, exp) => sum + parseFloat(exp.amount || 0), 0);
-      
-      const expenseUpdateData = {
-        firstName: updatedData.firstName,
-        lastName: updatedData.lastName,
-        departmentId: parseInt(updatedData.selectedDepartment),
-        date: updatedData.selectedDate || expenseDate.toISOString().split('T')[0],
-        totalAmount: totalAmount,
-        ExpenseItems: updatedData.expenses.map(exp => ({
-          id: exp.id,
-          paidTo: exp.paidTo,
-          purpose: exp.purpose,
-          categoryId: parseInt(exp.categoryId) || null,
-          amount: parseFloat(exp.amount),
-          status: exp.status || 'pending'
-        })),
-        userId: user.userId
-      };
+    // Small delay to prevent immediate closing when menu opens
+    const timer = setTimeout(() => {
+      document.addEventListener('click', handleClickOutside);
+    }, 0);
 
-      const expenseId = selectedExpense?.expenseId || selectedExpense?.id;
-      if (!expenseId) {
-        throw new Error('No expense ID found for updating');
-      }
-
-      await expenseAPI.updateExpense(expenseId, expenseUpdateData);
-      
-      await refetchExpenseData(expenseDate);
-      
-      queryClient.invalidateQueries({ 
-        queryKey: ['expenses', expenseDate] 
-      });
-      
-      toast.success('Expense updated successfully');
-      closeExpenseSummary();
-    } catch (error) {
-      console.error('Failed to update expense:', error);
-      toast.error(`Failed to update expense: ${error.message || 'Unknown error'}`);
-    }
-  };
+    return () => {
+      clearTimeout(timer);
+      document.removeEventListener('click', handleClickOutside);
+    };
+  }, [openMenuId]);
 
   if (isAuthenticating) {
     return null;
@@ -429,16 +283,6 @@ const Transaction = () => {
     handleDropdownClick
   };
 
-  const expenseHandlers = {
-    handleEditClick,
-    handleCancelClick,
-    handleEditChange,
-    handleSaveClick,
-    handleCancelInlineEdit,
-    toggleExpenseMenu: toggleExpenseMenuFixed, 
-    handleDropdownClick
-  };
-
   const summaryHandlers = {
     handleSummaryInputChange,
     handleMcNoChange,
@@ -452,12 +296,107 @@ const Transaction = () => {
     refundAmounts
   };
 
-  const refundDisplay = (
-    <td className="bg-gray-100 text-red-600 font-medium py-1 px-4 md:px-8 border border-gray-300 text-right">
-      {totalRefundsToDisplay.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}
-    </td>
-  );
-  
+  // Pagination calculations
+  const totalPages = Math.ceil(filteredTransactions.length / itemsPerPage);
+  const indexOfLastItem = currentPage * itemsPerPage;
+  const indexOfFirstItem = indexOfLastItem - itemsPerPage;
+  const currentTransactions = filteredTransactions.slice(indexOfFirstItem, indexOfLastItem);
+
+  const handlePageChange = (pageNumber) => {
+    setCurrentPage(pageNumber);
+    window.scrollTo({ top: 0, behavior: 'smooth' });
+  };
+
+  const renderPageNumbers = () => {
+    const pageNumbers = [];
+    const maxVisiblePages = 5;
+
+    if (totalPages <= maxVisiblePages) {
+      // Show all pages if total pages is less than or equal to maxVisiblePages
+      for (let i = 1; i <= totalPages; i++) {
+        pageNumbers.push(
+          <button
+            key={i}
+            onClick={() => handlePageChange(i)}
+            className={`px-3 py-1 rounded text-sm md:text-base ${
+              currentPage === i
+                ? 'bg-green-600 text-white'
+                : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+            }`}
+          >
+            {i}
+          </button>
+        );
+      }
+    } else {
+      // Show first page
+      pageNumbers.push(
+        <button
+          key={1}
+          onClick={() => handlePageChange(1)}
+          className={`px-3 py-1 rounded text-sm md:text-base ${
+            currentPage === 1
+              ? 'bg-green-600 text-white'
+              : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+          }`}
+        >
+          1
+        </button>
+      );
+
+      // Show dots if current page is far from start
+      if (currentPage > 3) {
+        pageNumbers.push(
+          <span key="dots-start" className="text-gray-500">...</span>
+        );
+      }
+
+      // Show pages around current page
+      const startPage = Math.max(2, currentPage - 1);
+      const endPage = Math.min(totalPages - 1, currentPage + 1);
+
+      for (let i = startPage; i <= endPage; i++) {
+        pageNumbers.push(
+          <button
+            key={i}
+            onClick={() => handlePageChange(i)}
+            className={`px-3 py-1 rounded text-sm md:text-base ${
+              currentPage === i
+                ? 'bg-green-600 text-white'
+                : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+            }`}
+          >
+            {i}
+          </button>
+        );
+      }
+
+      // Show dots if current page is far from end
+      if (currentPage < totalPages - 2) {
+        pageNumbers.push(
+          <span key="dots-end" className="text-gray-500">...</span>
+        );
+      }
+
+      // Show last page
+      pageNumbers.push(
+        <button
+          key={totalPages}
+          onClick={() => handlePageChange(totalPages)}
+          className={`px-3 py-1 rounded text-sm md:text-base ${
+            currentPage === totalPages
+              ? 'bg-green-600 text-white'
+              : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+          }`}
+        >
+          {totalPages}
+        </button>
+      );
+    }
+
+    return pageNumbers;
+  };
+
   return (
     <div className='flex flex-col md:flex-row min-h-screen bg-gray-100'>
       <div className="md:block md:w-64 flex-shrink-0">
@@ -466,50 +405,107 @@ const Transaction = () => {
       
       {/* Main content */}
       <div className="flex-grow p-2 md:p-4">
-        {/* Income section */}
-        <div className="bg-white rounded-lg shadow p-3 md:p-4 mb-4">
-          {/* Income header */}
-          <div className="flex flex-col md:flex-row justify-between items-center mb-3 md:mb-4">
-            <div className="flex items-center mb-3 md:mb-0">
-              <h2 className="text-xl md:text-2xl font-bold text-green-800 flex items-center">
-                Income
-                <span className="ml-2">
-                  <img src={Income} className="w-7 h-7 md:w-10 md:h-10" alt="Income Icon"/>
-                </span>
-              </h2>
+        {/* Header with title and date */}
+        <div className="bg-white rounded-lg shadow p-3 md:p-6 mb-4">
+          <div className="flex flex-col md:flex-row justify-between items-start md:items-center mb-4">
+            <h1 className="text-2xl md:text-3xl font-bold text-gray-800 mb-3 md:mb-0">Transactions</h1>
+            <div className="flex items-center space-x-2">
+              <span className="text-gray-600 text-sm md:text-base">Showing data for:</span>
+              <DateSelector 
+                date={selectedDate}
+                onDateChange={handleDateChange}
+                inputRef={incomeDateInputRef}
+              />
             </div>
-            
-            <div className="flex flex-col md:flex-row space-y-2 md:space-y-0 md:space-x-2 w-full md:w-auto">
-              <div className="relative w-full md:w-auto">
-                <input
-                  type="text"
-                  placeholder="Search Patient..."
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                  className="border-2 border-green-800 focus:border-green-800 focus:outline-none rounded-lg px-2 py-1 md:px-4 md:py-2 w-full text-sm md:text-base"
-                />
-                <div className="absolute right-2 top-1/2 transform -translate-y-1/2 text-gray-400">
-                  <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" className="md:w-4 md:h-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><circle cx="11" cy="11" r="8"></circle><line x1="21" y1="21" x2="16.65" y2="16.65"></line></svg>
+          </div>
+
+          {/* Summary Cards */}
+          <div className="grid grid-cols-2 md:grid-cols-4 gap-3 md:gap-4">
+            {/* On-hand Card */}
+            <div className="bg-purple-600 rounded-lg p-4 md:p-6 text-white shadow-md">
+              <div className="flex items-center justify-between mb-2">
+                <div className="bg-white bg-opacity-20 rounded-full p-2">
+                  <svg className="w-5 h-5 md:w-6 md:h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M12 8c-1.657 0-3 .895-3 2s1.343 2 3 2 3 .895 3 2-1.343 2-3 2m0-8c1.11 0 2.08.402 2.599 1M12 8V7m0 1v8m0 0v1m0-1c-1.11 0-2.08-.402-2.599-1M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+                  </svg>
                 </div>
               </div>
-              
-              <div className="flex flex-col md:flex-row space-y-2 md:space-y-0 md:space-x-2 w-full md:w-auto">
-                <DateSelector 
-                  date={selectedDate}
-                  onDateChange={handleDateChange}
-                  inputRef={incomeDateInputRef}
-                />
-                    
-                <button onClick={handleNewIncome} className="px-3 md:px-8 py-1 md:py-2 bg-green-800 text-white rounded-md text-sm md:text-base flex-1 md:flex-none md:w-32 hover:bg-green-600">
-                  New
-                </button>
+              <div className="text-xs md:text-sm opacity-90 mb-1">On-hand</div>
+              <div className="text-xl md:text-2xl font-bold">₱ {depositAmount.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}</div>
+            </div>
+
+            {/* GCash Card */}
+            <div className="bg-blue-500 rounded-lg p-4 md:p-6 text-white shadow-md">
+              <div className="flex items-center justify-between mb-2">
+                <div className="bg-white bg-opacity-20 rounded-full p-2">
+                  <svg className="w-5 h-5 md:w-6 md:h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M3 10h18M7 15h1m4 0h1m-7 4h12a3 3 0 003-3V8a3 3 0 00-3-3H6a3 3 0 00-3 3v8a3 3 0 003 3z" />
+                  </svg>
+                </div>
+              </div>
+              <div className="text-xs md:text-sm opacity-90 mb-1">GCash</div>
+              <div className="text-xl md:text-2xl font-bold">₱ {totalGCash.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}</div>
+            </div>
+
+            {/* Refund Card */}
+            <div className="bg-orange-500 rounded-lg p-4 md:p-6 text-white shadow-md">
+              <div className="flex items-center justify-between mb-2">
+                <div className="bg-white bg-opacity-20 rounded-full p-2">
+                  <svg className="w-5 h-5 md:w-6 md:h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M16 15v-1a4 4 0 00-4-4H8m0 0l3 3m-3-3l3-3m9 14V5a2 2 0 00-2-2H6a2 2 0 00-2 2v16l4-2 4 2 4-2 4 2z" />
+                  </svg>
+                </div>
+              </div>
+              <div className="text-xs md:text-sm opacity-90 mb-1">Refund</div>
+              <div className="text-xl md:text-2xl font-bold">₱ {totalRefundsToDisplay.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}</div>
+            </div>
+
+            {/* Total Card */}
+            <div className="bg-green-600 rounded-lg p-4 md:p-6 text-white shadow-md">
+              <div className="flex items-center justify-between mb-2">
+                <div className="bg-white bg-opacity-20 rounded-full p-2">
+                  <svg className="w-5 h-5 md:w-6 md:h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M12 8c-1.657 0-3 .895-3 2s1.343 2 3 2 3 .895 3 2-1.343 2-3 2m0-8c1.11 0 2.08.402 2.599 1M12 8V7m0 1v8m0 0v1m0-1c-1.11 0-2.08-.402-2.599-1M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+                  </svg>
+                </div>
+              </div>
+              <div className="text-xs md:text-sm opacity-90 mb-1">Total</div>
+              <div className="text-xl md:text-2xl font-bold">₱ {totalAmount.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}</div>
+            </div>
+          </div>
+        </div>
+
+        {/* Transactions Table Section */}
+        <div className="bg-white rounded-lg shadow p-3 md:p-4">
+          {/* Search and Add Button */}
+          <div className="flex flex-col md:flex-row justify-between items-center mb-3 md:mb-4">
+            <div className="relative w-full md:w-96 mb-3 md:mb-0">
+              <input
+                type="text"
+                placeholder="Search..."
+                value={searchTerm}
+                onChange={(e) => setSearchTerm(e.target.value)}
+                className="border-2 border-gray-300 focus:border-green-600 focus:outline-none rounded-lg px-4 py-2 w-full text-sm md:text-base pr-10"
+              />
+              <div className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-400">
+                <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+                  <circle cx="11" cy="11" r="8"></circle>
+                  <line x1="21" y1="21" x2="16.65" y2="16.65"></line>
+                </svg>
               </div>
             </div>
+            
+            <button 
+              onClick={handleNewIncome} 
+              className="px-6 md:px-8 py-2 bg-green-600 text-white rounded-md text-sm md:text-base hover:bg-green-700 transition-colors w-full md:w-auto"
+            >
+              Add New
+            </button>
           </div>
           
           {/* Income Table */}
           <IncomeTable
-            filteredTransactions={filteredTransactions}
+            filteredTransactions={currentTransactions}
             departmentsWithValues={departmentsWithValues}
             departmentTotals={calculatedTotals.departmentTotals}
             totalGross={totalGross}
@@ -520,128 +516,56 @@ const Transaction = () => {
             handlers={rowHandlers}
           />
           
-          <div className="mt-2 flex flex-col md:flex-row justify-between p-2">
-            <div className="flex flex-wrap items-center mb-4 md:mb-0">
-              {filteredTransactions.length > 0 && (
-                <button 
-                  onClick={handleGenerateReport}
-                  className="bg-green-800 text-white px-4 md:px-6 py-2 rounded flex items-center mb-2 md:mb-0 text-sm md:text-base hover:bg-green-600"
-                >
-                  Generate Report <Download className="ml-2 h-3 w-3 md:h-4 md:w-4" />
-                </button>
+          {/* Generate Report Button */}
+          <div className="mt-4 flex justify-between items-center">
+            <div className="text-sm text-gray-600">
+              {filteredTransactions.length > 0 ? (
+                <>Showing {indexOfFirstItem + 1}-{Math.min(indexOfLastItem, filteredTransactions.length)} of {filteredTransactions.length} transactions</>
+              ) : (
+                <>No transactions found</>
               )}
             </div>
-            
-            {/* Summary Box */}
-            <div className="mt-2 md:mt-0 border border-gray-300 w-full md:w-auto">
-              <table className="border-collapse w-full md:w-auto text-sm md:text-base">
-                <thead className='bg-yellow-500 text-gray-800 font-bold'>
-                  <tr>
-                    <th className="px-4 py-2 text-white" colSpan="2">INCOME SUMMARY</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td className="bg-blue-500 text-white font-medium py-1 px-2 md:px-4 border border-gray-300 text-center">
-                      GCASH
-                    </td>
-                    <td className="bg-gray-100 text-green-800 font-medium py-1 px-4 md:px-8 border border-gray-300 text-right">
-                      {totalGCash.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}
-                    </td>
-                  </tr>
-                  <tr>
-                    <td className="bg-red-500 text-white font-medium py-1 px-2 md:px-4 border border-gray-300 text-center">
-                      REFUND
-                    </td>
-                    {refundDisplay}
-                  </tr>
-                  <tr>
-                    <td className="bg-green-800 text-white font-medium py-1 px-2 md:px-4 border border-gray-300 text-center">
-                      DEPOSIT
-                    </td>
-                    <td className="bg-gray-100 text-green-800 font-medium py-1 px-4 md:px-8 border border-gray-300 text-right">
-                      {Math.max(0, totalGross - totalRefundsToDisplay - totalGCash).toLocaleString(undefined, { 
-                        minimumFractionDigits: 2, 
-                        maximumFractionDigits: 2 
-                      })}
-                    </td>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-          </div>
-        </div>
-        
-        {/* Expenses section */}
-        <div className="bg-white rounded-lg shadow p-3 md:p-4">
-          <div className="flex flex-col md:flex-row justify-between items-center mb-3 md:mb-4">
-            <div className="flex items-center mb-3 md:mb-0">
-              <h2 className="text-xl md:text-2xl font-bold text-green-800 flex items-center">
-                Expenses
-                <span className="ml-2">
-                  <img src={Expense} className="w-7 h-7 md:w-10 md:h-10" alt="Income Icon"/>
-                </span>
-              </h2>
-            </div>
-            
-            <div className="flex flex-col md:flex-row space-y-2 md:space-y-0 md:space-x-2 w-full md:w-auto">
-              <div className="relative w-full md:w-auto">
-                <input
-                  type="text"
-                  placeholder="Search Expenses..."
-                  value={expenseSearchTerm}
-                  onChange={(e) => setExpenseSearchTerm(e.target.value)}
-                  className="border-2 border-green-800 focus:border-green-800 focus:outline-none rounded-lg px-2 py-1 md:px-4 md:py-2 w-full text-sm md:text-base"
-                />
-                <div className="absolute right-2 top-1/2 transform -translate-y-1/2 text-gray-400">
-                  <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" className="md:w-4 md:h-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><circle cx="11" cy="11" r="8"></circle><line x1="21" y1="21" x2="16.65" y2="16.65"></line></svg>
-                </div>
-              </div>
-              
-              <div className="flex space-x-2 w-full md:w-auto justify-between md:justify-start">
-                <DateSelector 
-                  date={expenseDate}
-                  onDateChange={handleExpenseDateChange}
-                  inputRef={expenseDateInputRef}
-                />
-                    
-                <button onClick={handleNewExpenses} className="px-3 md:px-8 py-1 md:py-2 bg-green-800 text-white rounded-md text-sm md:text-base flex-1 md:flex-none md:w-32 hover:bg-green-600">
-                  New
-                </button>
-              </div>
-            </div>
-          </div>
-          
-          {/* Expenses table section */}
-          <ExpenseTable
-            filteredExpenses={filteredExpenses || []}
-            totalExpense={totalExpense || 0}
-            editingId={editingId}
-            editedExpense={editedTransaction}
-            openMenuId={null}
-            handlers={{
-              handleEditClick,
-              handleCancelClick,
-              handleSaveClick,
-              handleCancelInlineEdit,
-              toggleExpenseMenu: null
-            }}
-            expenseSearchTerm={expenseSearchTerm}
-            onEditExpense={handleEditExpense}
-            key={`expense-table-${expenseDate.toISOString().split('T')[0]}`} 
-          />
-
-          {/* Expense Generate Report Button */}
-          <div className="mt-2 flex justify-start p-2">
-            {filteredExpenses.length > 0 && (
+            {filteredTransactions.length > 0 && (
               <button 
-                onClick={handleGenerateExpenseReport}
-                className="bg-green-800 text-white px-4 md:px-6 py-2 rounded flex items-center text-sm md:text-base hover:bg-green-600"
+                onClick={handleGenerateReport}
+                className="bg-green-600 text-white px-4 md:px-6 py-2 rounded-md flex items-center text-sm md:text-base hover:bg-green-700 transition-colors"
               >
-                Generate Report <Download className="ml-2 h-3 w-3 md:h-4 md:w-4" />
+                <Download className="mr-2 h-4 w-4" />
+                Generate Report
               </button>
             )}
           </div>
+
+          {/* Pagination */}
+          {totalPages > 1 && (
+            <div className="flex justify-center items-center space-x-2 mt-4">
+              <button 
+                onClick={() => handlePageChange(currentPage - 1)}
+                disabled={currentPage === 1}
+                className={`px-3 py-1 rounded text-sm md:text-base ${
+                  currentPage === 1
+                    ? 'bg-gray-300 text-gray-500 cursor-not-allowed'
+                    : 'bg-green-600 text-white hover:bg-green-700'
+                }`}
+              >
+                Prev
+              </button>
+              
+              {renderPageNumbers()}
+              
+              <button 
+                onClick={() => handlePageChange(currentPage + 1)}
+                disabled={currentPage === totalPages}
+                className={`px-3 py-1 rounded text-sm md:text-base ${
+                  currentPage === totalPages
+                    ? 'bg-gray-300 text-gray-500 cursor-not-allowed'
+                    : 'bg-green-600 text-white hover:bg-green-700'
+                }`}
+              >
+                Next
+              </button>
+            </div>
+          )}
         </div>
       </div>
       
@@ -674,27 +598,6 @@ const Transaction = () => {
         isPending={mutations?.cancelTransaction?.isPending}
       />
 
-      {/* Expense Summary Modal */}
-      {isExpenseSummaryOpen && selectedExpense && (
-        <ExpenseSummaryModal
-          isOpen={isExpenseSummaryOpen}
-          onClose={closeExpenseSummary}
-          firstName={getExpenseModalData().firstName}
-          lastName={getExpenseModalData().lastName}
-          selectedDepartment={getExpenseModalData().selectedDepartment}
-          selectedDate={getExpenseModalData().selectedDate}
-          departments={departments}
-          categories={categories}
-          expenses={getExpenseModalData().expenses}
-          calculateTotal={calculateExpenseTotal}
-          onSave={handleSaveExpenseChanges}
-          isLoading={false}
-          isEditing={isEditingExpense}
-          onEnterEditMode={() => setIsEditingExpense(true)}
-          mode="edit"
-        />
-      )}
-
       {/* WebAuthn Authentication Modal */}
       <WebAuthModal
         isOpen={isModalOpen}
@@ -722,4 +625,4 @@ const Transaction = () => {
   );
 };
 
-export default Transaction;
+export default NewTransaction;
