+++ frontend/src/utils/chartDataTransformers.js
@@ -281,8 +281,8 @@ export const formatCurrency = (amount, currency = '₱') => {
   }
   
   return `${currency} ${amount.toLocaleString('en-US', {
-    minimumFractionDigits: 0,
-    maximumFractionDigits: 0
+    minimumFractionDigits: 2,
+    maximumFractionDigits: 2
   })}`;
 };
